tests improved
